Create QuesEasy1
In this program, the input string "Hello World" is passed to the 'length_of_last_word()' function. The function removes the trailing spaces and splits the string into a list of words. The length of the last word, which is "World", is then calculated and returned as the output of the function.