Create QuesHard2
In this program, the input string is "aacecaaa". The 'shortest_palindrome()' function iterates through the string from both ends and compares the characters. If they don't match, the function adds the character at the beginning of the string and repeats the process. The function returns the shortest palindrome that can be found by performing this transformation.